circ deps changes
